Fix health state not recovering after a occured failure
When an exception is raised inside the method `autoscale`, the health
endpoint indicates this by responding with status code "503". If
kube-aws-autoscaler recovers on a subsequent run, the health endpoint
still responds with status code "503".
These changes fix this behaviour by setting the global variable `Healthy`
to `True` after each run of autoscale if no exception occurred.